Store

+++ server/channels/store/sqlstore/store.go
@@ -264,7 +264,9 @@ func New(settings model.SqlSettings, logger mlog.LoggerIFace, metrics einterface
 	store.stores.propertyField = newPropertyFieldStore(store)
 	store.stores.propertyValue = newPropertyValueStore(store)
 
-	store.stores.preference.(*SqlPreferenceStore).deleteUnusedFeatures()
+	if !store.skipMigrations {
+		store.stores.preference.(*SqlPreferenceStore).deleteUnusedFeatures()
+	}
 
 	return store, nil
 }
@@ -637,7 +639,6 @@ func (ss *SqlStore) DoesTableExist(tableName string) bool {
 			`SELECT count(relname) FROM pg_class WHERE relname=$1`,
 			strings.ToLower(tableName),
 		)
-
 		if err != nil {
 			mlog.Fatal("Failed to check if table exists", mlog.Err(err))
 		}
@@ -656,7 +657,6 @@ func (ss *SqlStore) DoesTableExist(tableName string) bool {
 		    `,
 			tableName,
 		)
-
 		if err != nil {
 			mlog.Fatal("Failed to check if table exists", mlog.Err(err))
 		}
@@ -679,7 +679,6 @@ func (ss *SqlStore) DoesColumnExist(tableName string, columnName string) bool {
 			strings.ToLower(tableName),
 			strings.ToLower(columnName),
 		)
-
 		if err != nil {
 			if err.Error() == "pq: relation \""+strings.ToLower(tableName)+"\" does not exist" {
 				return false
@@ -703,7 +702,6 @@ func (ss *SqlStore) DoesColumnExist(tableName string, columnName string) bool {
 			tableName,
 			columnName,
 		)
-
 		if err != nil {
 			mlog.Fatal("Failed to check if column exists", mlog.Err(err))
 		}
@@ -725,7 +723,6 @@ func (ss *SqlStore) DoesTriggerExist(triggerName string) bool {
 			WHERE
 				tgname = $1
 		`, triggerName)
-
 		if err != nil {
 			mlog.Fatal("Failed to check if trigger exists", mlog.Err(err))
 		}
@@ -742,7 +739,6 @@ func (ss *SqlStore) DoesTriggerExist(triggerName string) bool {
 				trigger_schema = DATABASE()
 			AND	trigger_name = ?
 		`, triggerName)
-
 		if err != nil {
 			mlog.Fatal("Failed to check if trigger exists", mlog.Err(err))
 		}
